Fixed a signed comparison warning.

+++ test/serialization/detail/test_serialized.cpp
@@ -597,7 +597,7 @@ NNTestClassImpl(Serialized)
                 s.set(orig);
 
                 Storage<size_t> copy = s.get<Storage<size_t>>();
-                for(int i = 0; i < 6; ++i)
+                for(size_t i = 0; i < 6; ++i)
                     NNTestEquals(copy[i], i);
             }
